Update README.md
When executing the script ".py" could be omitted because of the shebang "#!/usr/bin/env python3" so that if the script is added as a command by exporting the path it could be run as `addspotter` instead of `adspotter.py`  but since you kept it in the file, i added it in the usage to be consistent. Eventually enhanced the installation commands markdown. I suggest you to add a description like "A Spotify adblocker for Linux" in the description section and under the README title.
Also keep in mind that if you use "python" instead of "python3", if someone have python2 set for the "python" command it won't work because the script is not python2 compliant, i didnt change this tho because it might be your preference.